Points
Points text has been added to the player UI. Player UI has been saved as a prefab.

Game object created and saved as a prefab with the score trigger script assigned to it. Assigns player points when they make contact with the object. A delay in points earned can also be configured so the player cannot keep earning points by touching object